added: ViT backbone is now available

+++ models/dab_conv_trans_detr/dab_transformer.py
@@ -326,7 +326,7 @@ class Transformer(nn.Module):
 
         # encoder
         with torch.autocast("cuda", dtype=torch.float16, enabled=False):
-            memory = self.encoder(src_flatten, spatio_temporal_shapes, level_start_index, valid_ratios, lvl_pos_embed_flatten, mask_flatten)
+            memory = self.encoder(src_flatten.float(), spatio_temporal_shapes, level_start_index, valid_ratios, lvl_pos_embed_flatten, mask_flatten)
         
         # revert to the original shape
         srcs_per_lvl = []
